+++ libsplinter/src/admin/rest_api/mod.rs
@@ -24,7 +24,7 @@ use crate::admin::service::AdminService;
 use crate::admin::store::AdminServiceStore;
 use crate::rest_api::actix_web_1::{Resource, RestResourceProvider};
 #[cfg(all(feature = "authorization", feature = "rest-api-actix"))]
-use crate::rest_api::auth::Permission;
+use crate::rest_api::auth::authorization::Permission;
 
 #[cfg(all(feature = "authorization", feature = "rest-api-actix"))]
 const CIRCUIT_READ_PERMISSION: Permission = Permission::Check("circuit.read");
