+++ libsplinter/src/biome/rest_api/actix/key_management.rs
@@ -24,7 +24,7 @@ use crate::biome::rest_api::resources::key_management::{NewKey, ResponseKey, Upd
 use crate::futures::{Future, IntoFuture};
 use crate::protocol;
 #[cfg(feature = "authorization")]
-use crate::rest_api::auth::Permission;
+use crate::rest_api::auth::authorization::Permission;
 use crate::rest_api::{
     actix_web_1::{into_bytes, HandlerFunction, Method, ProtocolVersionRangeGuard, Resource},
     ErrorResponse,
