+++ libsplinter/src/biome/rest_api/actix/logout.rs
@@ -23,7 +23,7 @@ use crate::biome::rest_api::{
 use crate::futures::IntoFuture;
 use crate::protocol;
 #[cfg(feature = "authorization")]
-use crate::rest_api::auth::Permission;
+use crate::rest_api::auth::authorization::Permission;
 use crate::rest_api::{
     actix_web_1::{HandlerFunction, Method, ProtocolVersionRangeGuard, Resource},
     secrets::SecretManager,
