+++ libsplinter/src/biome/rest_api/actix/verify.rs
@@ -18,7 +18,7 @@ use crate::actix_web::HttpResponse;
 use crate::futures::{Future, IntoFuture};
 use crate::protocol;
 #[cfg(feature = "authorization")]
-use crate::rest_api::auth::Permission;
+use crate::rest_api::auth::authorization::Permission;
 use crate::rest_api::{
     actix_web_1::{into_bytes, Method, ProtocolVersionRangeGuard, Resource},
     secrets::SecretManager,
