+++ libsplinter/src/oauth/rest_api/actix/callback.rs
@@ -34,7 +34,7 @@ use crate::oauth::{
 };
 use crate::protocol;
 #[cfg(feature = "authorization")]
-use crate::rest_api::auth::Permission;
+use crate::rest_api::auth::authorization::Permission;
 use crate::rest_api::{
     actix_web_1::{Method, ProtocolVersionRangeGuard, Resource},
     ErrorResponse,
