+++ libsplinter/src/oauth/rest_api/actix/login.rs
@@ -21,7 +21,7 @@ use std::collections::HashMap;
 use crate::oauth::OAuthClient;
 use crate::protocol;
 #[cfg(feature = "authorization")]
-use crate::rest_api::auth::Permission;
+use crate::rest_api::auth::authorization::Permission;
 use crate::rest_api::{
     actix_web_1::{Method, ProtocolVersionRangeGuard, Resource},
     ErrorResponse,
