+++ libsplinter/src/registry/rest_api/mod.rs
@@ -20,7 +20,7 @@ mod resources;
 
 use crate::rest_api::actix_web_1::{Resource, RestResourceProvider};
 #[cfg(all(feature = "authorization", feature = "rest-api-actix"))]
-use crate::rest_api::auth::Permission;
+use crate::rest_api::auth::authorization::Permission;
 
 use super::RwRegistry;
