+++ libsplinter/src/rest_api/auth/actix.rs
@@ -233,7 +233,7 @@ mod tests {
 
     use crate::error::InternalError;
     #[cfg(feature = "authorization")]
-    use crate::rest_api::auth::Permission;
+    use crate::rest_api::auth::authorization::Permission;
     use crate::rest_api::auth::{identity::Identity, AuthorizationHeader};
 
     /// Verifies that the authorization middleware sets the `Access-Control-Allow-Credentials: true`
