+++ libsplinter/src/rest_api/auth/authorization/permission_map.rs
@@ -12,7 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::rest_api::{auth::Permission, Method};
+use crate::rest_api::Method;
+
+use super::Permission;
 
 /// A map used to correlate requests with the permissions that guard them.
 #[derive(Default)]
